% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_se.R
\name{compute_se}
\alias{compute_se}
\title{Standard Errors (SE) wrapper for MMGSEM}
\usage{
se(object, d = 1e-03, naive = FALSE)
}
\arguments{
\item{object}{A fitted model returned by the mmgsem function.}

\item{d}{Step size for numerical approximation of the derivative. Default is 1e-03.}

\item{naive}{Logical. If TRUE, returns uncorrected SEs (faster but less accurate). If FALSE (default), applies the correction for step-wise estimation.}
}
\value{
SE_vector: List. Includes several vectors that contain the standard errors per type of parameter (beta, psi, etc.)
}
\description{
A wrapper to compute the SE of a fitted object by mmgsem(). It obtains the SE by numerically computing the second derivative of the loglikelihood function.
}
