% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Example data}
\format{
\subsection{}{

A data frame with 100800 rows and 21 columns:
\describe{
\item{V1-V20}{Observed variables of our example data}
\item{group}{Grouping variable}
}
}
}
\usage{
data_example
}
\description{
Example data used in the tutorial paper of mmgsem
}
\keyword{datasets}
