% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Model extraction function for MMGSEM}
\usage{
extract(object)
}
\arguments{
\item{object}{A resulting fitted object from the ModelSelection function.}

\item{nclus}{The number of clusters of the desired model}
}
\value{
selected_model: the selected model
}
\description{
Function used to extract an specific model out of a Model Selection (mmgsem) fitted object.
}
