% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{five_K_se}
\alias{five_K_se}
\title{Standard errors (5K model)}
\format{
\subsection{}{

List
\describe{
\item{SE_vector}{List of vectors containing the standard errors per type of parameter.}
\item{HESS}{Hessian matrix (second derivative) of the loglikelihood function with respect of all parameters.}
\item{vcov_betas}{Variance-covariance matrix of the beta parameters. Used for hypothesis testing.}
}
}
}
\usage{
five_K_se
}
\description{
List object containing the standard errors of the 5-cluster model from the example data of the tutorial
}
\keyword{datasets}
