% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{ModelSelection}
\alias{ModelSelection}
\title{Model selection wrapper for MMGSEM}
\usage{
ModelSelection(
  dat,
  S1 = NULL,
  S2 = NULL,
  group,
  nclus,
  seed = NULL,
  userStart = NULL,
  s1_fit = NULL,
  max_it = 10000L,
  nstarts = 20L,
  printing = FALSE,
  partition = "hard",
  endogenous_cov = TRUE,
  endo_group_specific = TRUE,
  sam_method = "local",
  meanstr = FALSE,
  rescaling = F,
  ...
)
}
\arguments{
\item{dat}{Data frame containing observed variables and a grouping variable..}

\item{S1}{Step 1 (measurement model) specification using lavaan syntax. Can be a list of strings determing the number of measurement blocks (e.g., one string for the MM of factor 1, and a second string for the MM of factor 2).
If s1_type = "mplus", S1 must be written in mplus syntax.}

\item{S2}{Step 2 (structural model) specification using lavaan syntax.}

\item{group}{Name of the grouping variable (as a character string).}

\item{nclus}{Integer vector of length two. Determines the range of cluster solutions to estimate. It must contain the lower and upper boundaries of the models the user wants to estimate. For instance, nclus = c(1, 6) estimates models from 1 to 6 clusters.}

\item{seed}{Optional. Random seed for replicable results.}

\item{userStart}{Optional. A user-defined cluster membership matrix (dimensions: number of groups × number of clusters) with binary values (1 or 0; 1 indicates cluster membership). To be used when the user has prior knowledge about the data. There must be only one 1 for each row. Skips random starts.
Example for 6 groups and 2 clusters:

\if{html}{\out{<div class="sourceCode">}}\preformatted{                              [,1] [,2]
                        [1,]    1    0
                        [2,]    1    0
                        [3,]    1    0
                        [4,]    0    1
                        [5,]    0    1
                        [6,]    0    1
}\if{html}{\out{</div>}}}

\item{s1_fit}{Optional. A fitted model object for Step 1. Can be from lavaan (CFA), blavaan (BCFA), or MplusAutomation (ML-CFA).}

\item{max_it}{Maximum number of iterations for Step 2 (default = 10000).}

\item{nstarts}{Number of random starts for Step 2 (default = 20)..}

\item{partition}{Initialisation partition for random starts: "hard" (default) or "soft".}

\item{endogenous_cov}{Logical. If TRUE (default), residual covariances among purely endogenous latent variables are estimated. If FALSE, they are fixed to 0 and only the residual variances are estimated.}

\item{endo_group_specific}{Logical. If TRUE (default), residual covariances are group- and cluster-specific (i.e., they are estimated for every group-cluster combination). If FALSE, they are cluster-specific (i.e., they are fixed to be equal across groups within a cluster). Note that, if FALSE, the residual covariances will also influence the clustering.}

\item{sam_method}{either "local" or "global. Follows local and global approaches from the SAM method. GLOBAL NOT FUNCTIONAL YET.}

\item{meanstr}{Logical. If TRUE, includes the mean structure in the model (e.g., for scalar invariance).}

\item{rescaling}{Only used when data is ordered. By default, MMGSEM uses the marker variable scaling approach. But identification
constraints with ordinal data (by default) are handled by standardizing the factors' variance in the first step.
The rescaling argument (either T or F) rescales the factor variances and loadings to the marker variable scaling
before running step 2. It is set to T by default (rescaling happens). If set to F, the factor variances are kept fixed to 1.}

\item{...}{MMGSEM relies on lavaan for the estimation of the first step (i.e., CFA). If needed, the users can pass any lavaan argument to MMGSEM
and it will be considered when estimating the CFA. For instance, std.lv if users want standardized latent variables,
group.equal for constraints, group.partial for non-invariances, etc.}

\item{s1_type}{String. Determines which model is used when estimating step 1. Can be "lavaan" (CFA), "blavaan" (BCFA), or "mplus" (ML-CFA).}

\item{ordinal}{Logical. If TRUE, observed variables are treated as ordinal (default = FALSE)}
}
\value{
Overview: Matrix. It contains the results of all the model selection measures for each model.

Models: List containing the separate mmgsem() results for each of the fitted models.
}
\description{
A wrapper to perform model selection of MMGSEM. It is used to answer the common clustering question: How many clusters should we use?
It calls the main function MMGSEM several times. One for each required model. Most of its arguments are the same as the MMGSEM main function.
}
