% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mmgsem.R
\name{plot.MMGSEM}
\alias{plot.MMGSEM}
\title{Plot model selection results}
\usage{
\method{plot}{MMGSEM}(ModelSel)
}
\arguments{
\item{ModelSel}{Object returned from the ModelSel() function.}
}
\value{
plots List that contains the plots for all the model selection measures.
}
\description{
Plots the graphs often used to determine the optimal number of clusters (e.g., scree plots). Provides plots for all the model selection measures
}
\details{
INPUT: Arguments required by the function
}
