% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_errors}
\alias{standard_errors}
\title{Standard errors (4K model)}
\format{
\subsection{}{

List
\describe{
\item{SE_vector}{List of vectors containing the standard errors per type of parameter.}
\item{HESS}{Hessian matrix (second derivative) of the loglikelihood function with respect of all parameters.}
\item{vcov_betas}{Variance-covariance matrix of the beta parameters. Used for hypothesis testing.}
}
}
}
\usage{
standard_errors
}
\description{
List object containing the standard errors of the 4-cluster model from the example data of the tutorial
}
\keyword{datasets}
