% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mmgsem.R
\name{summary.mmgsem}
\alias{summary.mmgsem}
\title{Summary function for mmgsem}
\usage{
\method{summary}{mmgsem}(model, se = NULL, model_selection = F)
}
\arguments{
\item{model}{Resulting object from a mmgsem model (must be a list). It can also be the resulting object from the ModelSelection() function.}

\item{se}{Must be the resulting object from the se() function of the mmgsem package. If included, the summary function will return the hypothesis testing of the relevant parameters (regressions).}
}
\description{
Summarizes the most important information of the results from a MMG-SEM model
}
