% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mmgsem.R
\name{test.mmgsem}
\alias{test.mmgsem}
\title{Summary function for mmgsem}
\usage{
test.mmgsem(model, se, multiple_comparison = FALSE)
}
\arguments{
\item{model}{Resulting object from a mmgsem model (must be a list). It can also be the resulting object from the ModelSelection() function.}

\item{se}{Must be the resulting object from the se() function of the mmgsem package. If included, the summary function will return the hypothesis testing of the relevant parameters (regressions).}
}
\description{
Summarizes the most important information of the results from a MMG-SEM model
}
