% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_se.R
\name{compute_se}
\alias{compute_se}
\title{Standard Errors (SE) wrapper for MMGSEM}
\usage{
se(object)
}
\arguments{
\item{object}{A resulting fitted object from the MMGSEM function.}
}
\value{
SE: List with the SE for all parameters. The SE are inside matrices mimicking the matrices of the parameters themselves. It also contains the corrected SE.

HESS: Hessian matrix containing all the second derivatives of the parameters (equivalent to the SE, but in other presentation).
}
\description{
A wrapper to compute the SE of a fitted object by MMG-SEM. It will use the Hessian matrix to compute the SE of both the measurement and the structural parameters.
}
