% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mmgsem.R
\name{plot.mmgsem}
\alias{plot.mmgsem}
\title{Plot model selection results}
\usage{
\method{plot}{mmgsem}(ModelSel, criteria)
}
\arguments{
\item{ModelSel}{Object returned from the ModelSel() function.}
}
\value{
plots List that contains the plots for all the model selection measures.
}
\description{
Plots the graphs often used to determine the optimal number of clusters (e.g., scree plots). Provides plots for all the model selection measures
}
\details{
INPUT: Arguments required by the function
}
